Move game structure dumper to it's own source
This could be used in game in some kind of diagnostic mode to dump the current values of in game structures.
