+++ src/Windows Code Release 3/Dump Scenario/CMakeLists.txt
@@ -1,5 +1,6 @@
 add_executable (DumpScenario
 	../endian_adjust.cpp
+	../dump_game_structures.cpp
 	dump_scenario.cpp
 	dump_scenario_cli.cpp
 )
