+++ src/Windows Code Release 3/Dump Scenario/dump_scenario.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
 #include <istream>
-#include <iterator>
+#include <ostream>
 
 bool dump_scenario(std::ostream_iterator<char> strm, std::istream& input);
